Add 'don't learn new cards today' option to custom study dialog
Add an option to avoid learning new cards today. This is useful when there are a lot cards to review and the user do not want to be overwhelmed by more new cards.

To achieve this result the new card limits are extended by a huge negative number.

A button to reset this behaviour, so allowing the user to learn new cards later in the day has been added.